refactor(diff): extract UnifiedDiffRows and SideBySideDiffRows
Move per-view-mode row rendering into dedicated subcomponents under
src/components/views/diff/. DiffView now delegates rendering to one of
the two based on viewMode.
